+++ lib/config/login_view_content.dart
@@ -2,17 +2,20 @@
 List<String> title =[
   "About Us",
   "Enthralling Events",
-  "Exciting Rewards"
+  "Workshops",
+  "Exciting Rewards",
 ];
 
 List<String> detail = [
   "Annual departmental fest of the department of Electrical Engineering, Prastuti provides curious minds a platform to lean, compete and develop their technical skills",
-  "The 21st iteration brings a series of 6 intellectual and innovative events with workshops and fun challenging problem statements",
-  "Tempeting Prizes to win through different competitions, Access to Guest Lectures by expert in the field, Excellent learning opportunities"
+  "The 21st iteration brings a series of may intellectual and innovative events with workshops and fun challenging problem statements",
+  "All workshops comprising organised content aim to serve as a vital point in the start of the participants' technological journey. Well curated resources would be provide to enhance every step of the way",
+  "Tempeting Prizes to win through different competitions, Access to Guest Lectures by expert in the field, Excellent learning opportunities",
 ];
 
 List<String> images = [
   "assets/login_view/awards.png",
   "assets/login_view/events.png",
   "assets/login_view/awards.png",
+  "assets/login_view/events.png",
 ];
