+++ lib/view_models/auth_view_model.dart
@@ -73,72 +73,3 @@ class AuthViewModelNotifier extends StateNotifier<bool>{
   }
 
 }
-
-Widget SignInButton(int, WidgetRef, BuildContext, bool) {
-  if (int < 2) {
-    return ElevatedButton(
-      onPressed: () {
-        null;
-      },
-      child: Image.asset('assets/login_view/google.png'),
-      style: ElevatedButton.styleFrom(
-        shape: const CircleBorder(),
-        backgroundColor: Colors.black,
-        fixedSize: Size(60, 60),
-        shadowColor: Colors.grey,
-        elevation: 15,
-      ),
-    );
-  } else {
-    return ElevatedButton(
-      onPressed: () {
-        WidgetRef.read(isLoggingIn.notifier).login(context: BuildContext);
-      },
-      child: bool
-          ? Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: [
-              Container(
-                height: 15,
-                width: 15,
-                child: CircularProgressIndicator(
-                    color: Colors.white,
-                  ),
-              ),
-              Text(
-                "Please Wait...",
-                style: TextStyle(
-                  fontSize: 15,
-                ),
-              )
-            ],
-          )
-          : Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                Container(
-                    height: 35,
-                    width: 35,
-                    child: Image.asset('assets/login_view/google.png')),
-                Container(
-                  width: 10,
-                ),
-                Text(
-                  "Log In",
-                  style: TextStyle(
-                    fontSize: 20,
-                    fontWeight: FontWeight.bold,
-                  ),
-                )
-              ],
-            ),
-      style: ElevatedButton.styleFrom(
-        shape: const StadiumBorder(),
-        backgroundColor: Colors.black,
-        fixedSize: Size(143, 50),
-        shadowColor: Colors.grey,
-        elevation: 15,
-      ),
-    );
-  }
-}
