+++ lib/views/authentication/login_view.dart
@@ -5,10 +5,7 @@ import 'package:prastuti_23/animations/login_view_animation.dart';
 import 'package:prastuti_23/config/screen_config.dart';
 import 'package:prastuti_23/view_models/auth_view_model.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
-import '../../animations/login_view_animation.dart';
 import '../../config/login_view_content.dart';
-import '../../config/screen_config.dart';
-import '../../view_models/auth_view_model.dart';
 
 class LoginView extends StatefulWidget {
   const LoginView({Key? key}) : super(key: key);
@@ -80,12 +77,16 @@ class _LoginViewState extends State<LoginView> with TickerProviderStateMixin {
                         mainAxisAlignment: MainAxisAlignment.center,
                         children: [
                           Obx(
-                            ()=> Padding(
-                              padding:  EdgeInsets.only(
-                                left: (1.0-animationController.pagePaddingValue.value)*100,
+                            () => Padding(
+                              padding: EdgeInsets.only(
+                                left: (1.0 -
+                                        animationController
+                                            .pagePaddingValue.value) *
+                                    100,
                               ),
                               child: Opacity(
-                                opacity: animationController.pagePaddingValue.value,
+                                opacity:
+                                    animationController.pagePaddingValue.value,
                                 child: Container(
                                   decoration: BoxDecoration(
                                     borderRadius: BorderRadius.circular(32),
@@ -108,11 +109,17 @@ class _LoginViewState extends State<LoginView> with TickerProviderStateMixin {
                           ),
                         ],
                       ),
-                      const SizedBox(height: 30,),
+                      const SizedBox(
+                        height: 30,
+                      ),
                       Obx(
-                        ()=> Padding(
-                          padding:  EdgeInsets.only(
-                                top: (1.0-animationController.pagePaddingValue.value)*75,),
+                        () => Padding(
+                          padding: EdgeInsets.only(
+                            top: (1.0 -
+                                    animationController
+                                        .pagePaddingValue.value) *
+                                75,
+                          ),
                           child: Opacity(
                             opacity: animationController.pagePaddingValue.value,
                             child: Column(
@@ -132,8 +139,10 @@ class _LoginViewState extends State<LoginView> with TickerProviderStateMixin {
                                   height: 15,
                                 ),
                                 SizedBox(
-                                  height: MediaQuery.of(context).size.height * 0.15,
-                                  width: MediaQuery.of(context).size.width * 0.80,
+                                  height:
+                                      MediaQuery.of(context).size.height * 0.15,
+                                  width:
+                                      MediaQuery.of(context).size.width * 0.80,
                                   child: Text(
                                     detail[index],
                                     style: const TextStyle(
@@ -150,8 +159,7 @@ class _LoginViewState extends State<LoginView> with TickerProviderStateMixin {
                           ),
                         ),
                       ),
-                    ]
-                ),
+                    ]),
               );
             },
           ),
@@ -165,16 +173,13 @@ class _LoginViewState extends State<LoginView> with TickerProviderStateMixin {
                 padding: const EdgeInsets.only(left: 15),
                 child: AnimatedSmoothIndicator(
                   activeIndex: _currentPage,
-                  count: 3,
+                  count: title.length,
                   effect: const WormEffect(
                     activeDotColor: Color(0xff272727),
                     dotHeight: 6.0,
                     dotWidth: 10,
                   ),
                 ),
-              ),
-              SizedBox(
-                width: SizeConfig.width * 0.43,
               ),
               Consumer(builder: (context, ref, child) {
                 bool isLoading = ref.watch(isLoggingIn);
@@ -185,5 +190,74 @@ class _LoginViewState extends State<LoginView> with TickerProviderStateMixin {
         )
       ],
     ));
+  }
+}
+
+Widget SignInButton(int, WidgetRef, BuildContext, bool) {
+  if (int < 3) {
+    return ElevatedButton(
+      onPressed: () {
+        null;
+      },
+      child: Image.asset('assets/login_view/google.png'),
+      style: ElevatedButton.styleFrom(
+        shape: const CircleBorder(),
+        backgroundColor: Colors.black,
+        fixedSize: Size(60, 60),
+        shadowColor: Colors.grey,
+        elevation: 10,
+      ),
+    );
+  } else {
+    return ElevatedButton(
+      onPressed: () {
+        WidgetRef.read(isLoggingIn.notifier).login(context: BuildContext);
+      },
+      child: bool
+          ? Row(
+              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+              children: [
+                Container(
+                  height: 15,
+                  width: 15,
+                  child: CircularProgressIndicator(
+                    color: Colors.white,
+                  ),
+                ),
+                Text(
+                  "Please Wait...",
+                  style: TextStyle(
+                    fontSize: 15,
+                  ),
+                )
+              ],
+            )
+          : Row(
+              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+              children: [
+                Container(
+                    height: 35,
+                    width: 35,
+                    child: Image.asset('assets/login_view/google.png')),
+                Container(
+                  width: 10,
+                ),
+                Text(
+                  "Log In",
+                  style: TextStyle(
+                    fontSize: 20,
+                    fontWeight: FontWeight.bold,
+                  ),
+                )
+              ],
+            ),
+      style: ElevatedButton.styleFrom(
+        shape: const StadiumBorder(),
+        backgroundColor: Colors.black,
+        fixedSize: Size(143, 50),
+        shadowColor: Colors.grey,
+        elevation: 10,
+      ),
+    );
   }
 }
